Another overhaul.
Config class now has accessors to commonly used parameters.
You can still get uncommon parameters using the 'get' and
'get_required' methods.

The ActionBatch and ActionFactory are now delegates of the
Pinto (via the 'enqueue' and 'create_action') methods.  This
creates the illusion of a simpler API.  And illusion is
good enough for me right now.