Minor overhaul, to improve performance.
We now have the noinit option, which basically
skips the checkout/update phase.  Don't use
this unless you know your working copy is sane.

The Store now keeps track of files that are added
or deleted.  This allows us to optimize and commit
only the files that have really changed, rather
than committing the whole wad, and expecting
svn (or git) to figure out what to do.