Major overhaul (again).
The Event tells the Store which files have been added or deleted.
In turn the Store only commits paths that have been added/deleted/modified.

After adding a local Dist, only the encumbent local Dist is removed from
the store (if there is one), not the mirrored Dist.  This way, when
you later decide to remove your local Dist entirely, the mirrored
one is immediately resurrected without having to mirror, which
would introduce even more changes to the repository that you may
not necessarily want.