Added lock mechanism & improved exception handling
ActionBatch now sets a lockfile before going to
work on the repository.  If we can't get a lock,
then an exception is thrown.  The lock is removed
when the work is done.

ActionBatch also traps and logs any exceptions that
arise during an action.  So if one action fails,
the process continues for the rest.