Major overhaul...
pinto-admin is now zero-conf:  The --repos argument
is required.  The remaining configuration settings
are stored in config/pinto.ini within the repository.
This allows you to have multiple repositories.

Pinto::Logger and Pinto::Config are now back in the
Pinto dist.  Now that these have a more specialized
purpose, we don't need to have them in a separate
dist.