Deal with really long arg lists.
When you first update your repository from a CPAN mirror,
you commit thousands of files.  On most systems, that
argument list is just too long.  So instead we use
a --targets file if the arg list is longer than 128.

128 is really just an arbitrary number.  The actual
limit is specified in bytes.  I'm not going to bother
figuring out the byte-size of the argument list, so
I just picked a number that should be within limits.