Added is_eligible_for_index flag.
This allows us to accept broken distributions (i.e. unsortable
or invalid versions) but exclude them from the index.  Since
this flag is stored in the database, we cannot easily change
the way it is computed.  This is unfortunate, but computing it
on the fly requires comparing it to ever other package to make
sure it is sortable.  We can probably minimize the headache by
offering a 'reindex' command like PAUSE.