A big refactoring.
The Pinto::Repository class is an abstraction of both the database
and the store.  This is where the primary methods for adding/deleting
archives will be.  This is also where I'll put convenience methods
that delegate to the db and store objects.

In most cases, the Action should only have to deal with the Repository
object.  But for occasional low-level stuff, you can still get to the
underlying db and store objects via the repository.