WARNING: Commit bomb!
Half-way through modest refactoring.  Here's the gist:

The IndexReader and IndexLoader are gone.  The IndexCache
now tells us what to mirror, since it already knows what
is in the index files of the remote repositories.

The Mirror action now does the heavy lifing to insert
records in the DB (this used to be done by IndexLoader).
As a benefit, distributions are fully mirrored one at a time,
rather than loading all the records in the DB, and *then*
mirroring each distribution.

Finally, created a DistributionSpec that should help with
creation of distribution objects in the DB.  Am now thinking
of moving most of the business logic into the Schema objects.