Vanquished the cleanup option.
This is just to hard to support with multiple repositories, because
a distribution might get added and then removed all in the same
batch.  I haven't figured out how to handle that in the Store
classes.  Presumably, they need to keep track of each add
and remove action, and then decide if that path really needs
to be comitted.  I bet Git would probably just do the right
thing, but Svn is more fickle.  So for now, I'm just going
to punt and say that auto-cleanup is not supported.  The next
best thing is to just run a cron job that runs the Clean
action periodically.