Pinto::Repository is now dumb.
It only knows how to do 2 things: add a dist, and remove a dist.
When I say "add" and "remove", I mean that it just tells the store
to add a file and puts some data in the database.

How that file gets there and what data is put into the database
is up to the Action.  This makes sense, because creating an Action
class is the primary means of extending Pinto.  So it has to
do most of the work of fetching a dist file, extracting its
packages, etc.