Convert versions to version objects.
When extracting packages or prerquisites from
a dist (known as pkg_specs), we now convert
the version into a proper version object.  This
makes it easy to compare them with each other.

However, versions are often invalid.  So you now
have a choice of whether to just toss out any
pkg_spec that has an invalid version, or if that
should be an error.