Slick improvements to Tester.
Instead of having to pass lots of arguments, you
can now specify test data with simple notation.

Also, simplified the API to just a couple main
assertions, that test whether a package is loaded.
These do all the work of inspecting the database
and filesystem to make sure everything is as it
should be.

Also added some convenient methods to assert whether
the log messages match some regex.  And added
a method to assert that the repository is
completely empty.

Not all the test scripts have been updated to conform
to the new API.  I'm about half way through.