Refactored the Store classes.
Pinto::Store is now an abstract class (of sorts).  It
provides the public API for adding/removing distributions.
But the mechanics of actually adding/removing files
is left to the subclasses.

Moreover, Pinto::Store provides the mandatory logic for
those methods (such as verifying that an added file
actually exists).  We then use inner() to call the
augmented method on the subclass, so it can do its
work.

Local storage on the filesystem (the default), is now
handled by a subclass, Pinto::Store::File.