Workaround for Subversion 1.7
Subversion 1.7 uses a different technique than previous
versions.  With 1.7, all the metadata is in a single .svn
directory at the top of the checkout, rather than in separate
.svn directories in each subdirectory of the project.

This means our technique of looking for a .svn folder to determine
the edges of the repository will no longer work.  Instead,
we need to actually ask Subversion if a file/directory is part
of the repository yet.

This is experimental, and I have not yet tried it on older Subversions.