Refactored guts into the Repository class
Now the Repository knows how to add distributions from whole-cloth
archives.  There are options to control whether or not to index
the archive and to recurse into the prereqs.

It makes sense to put this logic in the Repository, since the
add, import, and mirror actions all do this same fundamental
operation (just with different options).