Partially fleshed out the whole stack & pin stuff
Right now, only the Add and List actions are really
functional.  I need to tidy up the API on the Database
and Repository classes too.  For now, I'm thinking
that the Database API will only deal in real objects
and will have no argument validation.  In contrast,
the Repository API will deal in strings and things
and will validate arguments as needed.