BIG CHANGES: Stacks are now symlinks
Each stack is now represented on the filesystem with
its own 02packages file and a symlink back to the authors
dir.  This eliminates the need for Pinto::Server to generate
the 02packages dynamically, which makes it simpler and faster.
It also enables us to lock each stack separately, which
should prevent unecessary contention.