Eliminate the Operator role.
This simplifies things a bit.  But now all actions
happen in a transaction and will lock the repository.
I can probably improve the locking situation by allowing
some actions to require a shared lock instead of an
exclusive one.

But the whole idea behind the Operator role was to support
revision history, which I'm not sure how to do anyway.  So
until we really need it, lets get rid of it.