When asked to pull a module that is satisfied by the core, then skip it.
find_or_pull() also skips pulling any URL that looks like it is a perl
distribution.  But core-only modules aren't even in the CPAN index,
so we wouldn't get to the point where we have a URL to check.  So we
must determine if the module is already in core *before* we even call
find_or_pull().