Don't bother parsing the perl version.
I ran into an old version of version.pm which has this in its META.yml.

    requires:
      perl: '> 5.005'

This is valid meta spec, but an invalid version.  Work around it by bailing out
as early as possible on a Perl version.  The real fix is to read these versions
using CPAN::Meta.