Attempting to add a duplicate archive just causes the original to be re-registered.
In Schwern's use case, he was throwing lots of distributions into the repository
and didn't want to think about which ones he might have already added at an earlier
time.  So now, if you attempt to add a duplicate dist, Pinto just assumes that you're
really just trying to reindex what is already in the repository.  So that's what
it does.  But you'll get a warning to let you know that is going on.