The file path and checksum are no longer mandatory for packages
Some distributions generate the packages during configuration,
so it would be impossible to know the file path and checksum
without actually doing the configuration.  Even then, it
might be totally dependent on the environment you're in.

So we'll just make them optional.  If we can know the path
and checksum, then great.  But if not, then we'll live without.