Stack properties are now stored in a serialized JSON hash.
This idea was inspired by conversations with Josh Berkus
and David Fetter.  PostgreSQL has an HStore data type, which
can do some interesting things.  But SQLite has no such animal,
so we achieve something similar by using JSON to serialize the
hash into text.

Storing the props in this way is simple and convenient.  The
downside is that you can't really query on the properties or
impose constraints.  But that's probably ok because stack
properties don't really have any relationship to other fields
in the database.  Not being able to query is a slight bummer,
but I can probably live with that since the total number of
stacks is likely to be small (i.e less than 100).