Renamed allow_overlapping_distributions to prohibit_partial_distributions.
So the default beahvior will be like PAUSE.  To get the behavior
Schwern wants in #48, you have to enable the stack property.

I'm not really ready to commit to this feature yet, so I'm leaving
it undocumented for now.  It's not even set in the default properties
for the stack.  So you have to know the code (or Schwern) to enable
this feature.