Finish off the work to make the Action stop when the connection is lost.
So yes, it seems well known that the only way to know of a TCP
connection is broken is to actually write some data and see if
it succeeds.  Of course, this was all news to me.  So now we
have a special null message that we can send and the client will
just ignore.