Stop collecting package file paths and sha digests.
For many CPAN distributions, that information is
wrong anyway, which causes Dist::Metadata to blow
up when we try to get the digest of a path that
doesn't exist.  We're not really using this info
so I'm going to stop collecting it.  Eventually,
I want to take it out of the schema entirely.