Hacking environment for Pinto
The added script eases setting up an environment for hacking Pinto.

The assumption is that the user checks out Pinto's repository from
GitHub and then runs:

   etc/hack.sh

This creates a local library with relevant modules from Stratopan
and puts them in subdirectory llib. It also adds a bash file that
can be source-d to setup the right environment variables.

Apart from the script, there is a modification to .gitignore to
ignore the llib subdirectory.

Before merging, it might be considered to change "llib" into
something more intuitive like "hacklib".