Experiment with an option to control recursion behavior.
This is a repository setting.  Not sure if it makes sense
to also have this as a stack setting.

This creates an incompatible change in the pinto command-line
interface.

It also creates an inconsistency in the way we name command-line
options.  I may change all the others to remain consistent.