Start handling prereqs with CPAN::Meta::Requirements.
This allows prereqs to be specified as things like:

>= 1.2, < 2.4, != 1.6

A handful of distributions already declare prereqs
like this, but Pinto could not index them because
it expected prereqs to be simple version numbers.
Now we can index them.

More importantly, this paves the way toward specifying
target packages at the command line.  So you can
specify exactly which version of package you want to
pull.