Require newer Module::CoreList.
This is a hack!  If you happen to have installed Module::CoreList
separately and it is older than your version of Perl, then it
won't have entries for your Perl.  That causes coercion of the PerlVersion
type to fail.  So we now require the latest version of Module::CoreList
(at this moment, it is 3.02).

The right solution is to dynamically configure the build to inspect
your Module::CoreList and figure out if it is at least as new as your
Perl.  If not, then an upgrade to latest is needed.  But I haven't
figured out how to do dynamic prereqs with Dist::Zilla.