Only put simile packages in the index.
This simiple change solves a whole bunch of indexing headaches.
The problem is that some distributions will monkey patch others.
They usually do this by declaring the monkey-patched package in
a file that doesn't match the package name.  PAUSE will will not
try to index such packages, since it realizes the package actually
is from another distribution, and PAUSE *knows* every distribution
that has ever been released.

But Pinto can't know this, because it only *knows* about the
distributions it has seen.  As a result, Pinto ends up indexing
packages in distributions that it should not.  By itself, that
isn't too bad because you could always re-pull the original
distribution to correct the index.  But now that Pinto wants to
prohibit intermingled distributions, this solution is no longer
workable becasue you can only have one distribution or the other
in the index.

Now here's the kicker:  Remember that Pinto doesn't promise to
index like PAUSE anyway.  The index only has to be good enough
to install modules.  That means the index really only has to list
packages that can be use'd or require'd.  And that means that the
index only needs to have packages that are similes.  And that
means we can completely ignore all those monkey-patch packages
that are not similes.

The downside is that Pinto no longer has a true package index
like PAUSE.  But I think that's going to be ok.  In the vast
majority of cases, it only needs to index things that are
going to be used as dependencies.