An optimization that makes registration faster.
In repositories that prohibit intermingling, we
can assume that all packages in a distribution will
be registered.  So if we try to register a package
and find that it is already registered, then we
can assume that all the other packages in that
distribution have already been registered.  This
makes it much faster to re-pull packages.