Support path preservation in the root URI
The "root" attribute of Pinto::Root and Pinto::Root::Action is now treated
with a bit more care to not clobber the path and port components of the URI.
This allows Pinto::Server to operate at a URI with a path, such as:
http://hostname/my/perl/infrastructure/pinto

In order to test this, some refactoring of 01-requests.t was done, creating
a function to call for the test cases defined in a data structure. To rein
in the size of this structure a defaults hash is merged before the case is
tested.