Fix regression.
Need to find more sensible way to fetch a distribution
by author/archive from a particular stack. The trouble
is that the record appears many times for each package.
So we either need to GROUP BY, or find some other query
strategy.  For now, I'm cheating and just pulling the
first matching record.