Stop setting the default scheme and port.
The scheme is effectively mandatory anyway, since ```pinto```
requires it to distinguish between remote and local repositories.

And now, if you don't specify the port you'll get whatever is
appropriate for the scheme (e.g. HTTP=80, HTTPS=443).

This avoids lots of awkwardness about tring to find a sensible
default. This changes the interface, but it is more consistent
with how other applications work.  Closes GH #136.