Add a --cpanfile opt to the pull command.
Add a --cpanfile option to the pull command.  Option currently
requires a value.  TODO supply a default value of 'cpanfile'.

I had to "enhance" the way that arg_attribute and args_from_stdin
work.  Prior to this change they had no information about the context
in which they were running and just returned constants.

Now they are passed the $opts and $args from above and can decide
whether the Action should collect stuff from stdin or not.

In this case, if a cpanfile is specified then there should be no
additional args, either on the command line or on stdin.