Experiment to fix the "sticky wicket" for @hartzell
Installers are often asked to install a core module.
To make that work, the index of the pinto repository
needs to include core modules so the installer can
decide if the local version is up to date.

So, we inject artifical records into the index for all
the core modules as-of the target_perl_version.  In
theory, an installer should never actually try to
install these modules because it should already have
them.  But that assumes that the target_perl_version
actually is the same version of perl that is running
the installer.

This seems really icky, but it just might work.