Permit the root to be specified as an argument to the init command.
This helps people like @hartzell who are more accustomed to git.
But it breaks the interface -- now you must use the --stack option
to specify the name of the initial stack, instead of passing it
as an argument.  This probably makes sense, because you rarely
ever set the initial stack name, but you always have to specify
the root.  So you might as well make it easy to do so.