Workaround for funky dists.  Fixes GH #194.
* Improved the existing workarounds for common::sense and FCGI.
Now, we record the proper file name.  This will be useful in the
future when/if we extract docs from distributions.

* Added a new workaround for Net::LibIDN which generates the
true .pm file from _LibIDN.pm at build time.  I looked through
all the dists on CPAN and this seems to be the only one that
behaves this way.