Setting default action on deployable export format
It is now possible to set which action is default when calling a
generated deployable script via the export sub-command. The default
is to show a help message with further instructions, but this
option allows setting a different default (expecially 'install') to
generate script that should be "just run" (e.g. unattended).