Rate limit the download of upstream CHECKSUMS files
Normally the connection rate is tempered by the time it takes to unpack and analyse
a distribution.  The verify command runs a little faster, so insert a 500ms delay
between each verification by default.  The exact delay can be set via a --nice
option. Appologies for overloading cpan.stratopan.com :-(