Improve target_perl_version documentation
Module::CoreList is used to validate the Perl version number supplied to
the target_perl_version (tpv) property when creating a new stack.
Module::CoreList must be updated if you need to use a version of Perl
that is later than the ones it supports.  This revision clarifies that
it is Pinto's copy of Module::CoreList that must be updated.  It also
adds this information to the associated error message.