Fix indexing of modules that declare a provides file of META.{json,yml}
This treats dists that provide a package file of META.yml or META.json
as simile's.  This META.yml or .json is perfectly valid per the CPAN
Meta Spec, but Pinto was unable to deal with it.

This fixes pulling Term::ReadKey >= v2.37, and probably other modules as
well.  As a result, the Term::ReadKey specific workaround has been
removed

Added test case for pulling a dist where the provides file is either
META.yml or META.json

Fixes #240
Fixes #241